Refactor layout_information

+++ appcui/src/controls/layout/alignament.rs
@@ -1,4 +1,5 @@
 #[repr(u8)]
+#[derive(Copy,Clone)]
 pub(super) enum Alignament {
     TopLeft = 0,
     Top,
@@ -209,7 +210,7 @@ static HASH_COLISION_VALIDATOR: [u64; 95] = [
 
 impl Alignament {
     pub(super) fn new(hash: u64) -> Option<Alignament> {
-        let entry_index = hash % 95;
+        let entry_index = (hash % 95) as usize;
         if HASH_COLISION_VALIDATOR[entry_index] != hash {
             return None;
         }
